Match s03d_800D3ACC via overlay3.c merge (cross-object jump table)
