--- asm/overlays/s03d/s03d_800D3ACC.s
@@ -1,41 +1,0 @@
-	opt	c+, at+, e+, n-
-	section .text
-
-	xdef s03d_800D3ACC
-s03d_800D3ACC:
-	dw 0x27BDFFE8 ; 800D3ACC
-	dw 0xAFBF0010 ; 800D3AD0
-	dw 0x84830ACA ; 800D3AD4
-	dw 0x00000000 ; 800D3AD8
-	dw 0x2C620005 ; 800D3ADC
-	dw 0x1040001A ; 800D3AE0
-	dw 0x3C02800E ; 800D3AE4
-	dw 0x2442BB18 ; 800D3AE8
-	dw 0x00031880 ; 800D3AEC
-	dw 0x00621821 ; 800D3AF0
-	dw 0x8C620000 ; 800D3AF4
-	dw 0x00000000 ; 800D3AF8
-	dw 0x00400008 ; 800D3AFC
-	dw 0x00000000 ; 800D3B00
-	dw 0x0C034DC9 ; 800D3B04
-	dw 0x00000000 ; 800D3B08
-	dw 0x08034ED3 ; 800D3B0C
-	dw 0x00000000 ; 800D3B10
-	dw 0x0C034E61 ; 800D3B14
-	dw 0x00000000 ; 800D3B18
-	dw 0x08034ED3 ; 800D3B1C
-	dw 0x00000000 ; 800D3B20
-	dw 0x0C034E12 ; 800D3B24
-	dw 0x00000000 ; 800D3B28
-	dw 0x08034ED3 ; 800D3B2C
-	dw 0x00000000 ; 800D3B30
-	dw 0x0C034E9F ; 800D3B34
-	dw 0x00000000 ; 800D3B38
-	dw 0x08034ED3 ; 800D3B3C
-	dw 0x00000000 ; 800D3B40
-	dw 0x0C034DD5 ; 800D3B44
-	dw 0x00000000 ; 800D3B48
-	dw 0x8FBF0010 ; 800D3B4C
-	dw 0x00000000 ; 800D3B50
-	dw 0x03E00008 ; 800D3B54
-	dw 0x27BD0018 ; 800D3B58
